refactor: 투두 상태 관련 컴포넌트 로직 분리하여 dumb 하게 개선 [TM-62]
