+++ src/entities/todo/ui/TodoStatus.tsx
@@ -1,44 +1,22 @@
-import { ButtonHTMLAttributes, useState } from 'react';
+import { ButtonHTMLAttributes } from 'react';
 import styled, { keyframes } from 'styled-components';
 import CheckIcon from '@/shared/assets/todo/check.svg?react';
 import { Status } from '../todo.type';
-import useTodoQuries from '@/features/todo/model/useTodoQuries';
-import { STATUS_TO_OPTION } from '../constants/todo';
 
 interface ITodoStatus extends ButtonHTMLAttributes<HTMLButtonElement> {
-  status: Status;
+  todoStatus: Status;
   modalToggle: () => void;
   todoId: number;
 }
 
-function TodoStatus({ status, todoId, modalToggle, ...props }: ITodoStatus) {
-  const { useEditTodoStatusMutation } = useTodoQuries();
-  const { mutate: editTodoStatus } = useEditTodoStatusMutation(todoId);
-
-  const [todoStatus, setTodoStatus] = useState<Status>(status);
-
-  // 관련 로직 분리 필요
-  const handleTodoStatus = () => {
-    const option = STATUS_TO_OPTION[status]; // 상태에 따른 option 값 가져오기
-    editTodoStatus(option); // 상태 변경 API 호출
-
-    // ui 상태 변경
-    if (todoStatus === 'PENDING') {
-      setTodoStatus('IN_PROGRESS');
-    } else if (todoStatus === 'IN_PROGRESS') {
-      setTodoStatus('COMPLETED');
-      modalToggle(); // 이미지 업로드 모달 열기
-    } else if (todoStatus === 'COMPLETED') {
-      setTodoStatus('PENDING');
-    }
-  };
-
+function TodoStatus({
+  todoStatus,
+  todoId,
+  modalToggle,
+  ...props
+}: ITodoStatus) {
   return (
-    <TodoStatusButton
-      $todoStatus={todoStatus}
-      {...props}
-      onClick={handleTodoStatus}
-    >
+    <TodoStatusButton $todoStatus={todoStatus} {...props}>
       <IconWrapper $todoStatus={todoStatus}>
         {todoStatus === 'IN_PROGRESS' || <CheckIcon />}
