+++ src/entities/todo/ui/index.ts
@@ -1,5 +1,4 @@
-export * from '../../../features/todo/ui/Todo';
 export * from './Button';
-export * from './TodoStatus';
 export * from './ProgressStatusBadge';
 export * from './ProgressBar';
+export * from './TodoStatus';
